+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Comment and reply counts on the video details page now update in real time when comments and replies are fetched.
+  (mddburgess/winston#43)
+
 ## [1.2.1] — 2025-04-16
 
 ### Fixed
@@ -72,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Web UI to fetch channels, videos, and comments.
 - Cache channel and video thumbnails.
 
+[Unreleased]: https://github.com/mddburgess/winson/compare/v1.2.1...HEAD
 [1.2.1]: https://github.com/mddburgess/winston/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/mddburgess/winston/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/mddburgess/winston/compare/v1.0.0...v1.1.0